# backend/docs/swagger.yaml
basePath: /
definitions:
  controllers.AuthRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  controllers.AuthResponse:
    properties:
      email:
        type: string
      first_session:
        type: boolean
      id:
        type: integer
      token:
        type: string
    type: object
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controllers.ResetPasswordRequest:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - code
    - email
    - new_password
    type: object
...
  controllers.VerifyRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      first_session:
        type: boolean
      id:
        type: integer
      is_deleted:
        type: boolean
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      updated_at:
        type: string
      verification_code:
        type: string
      verification_code_expiry:
        type: string
      verified:
        type: boolean
    required:
    - email
    - first_name
    - password
    type: object
...
host: localhost:5050
info:
  contact:
    email: support@swagger.io
    name: Arly API Support
    url: http://www.swagger.io/support
  description: This is the Arly API server documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Arly API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by checking their email, password, and verification
        status.
      parameters:
      - description: User login data
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/controllers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User authenticated successfully
          schema:
            $ref: '#/definitions/controllers.AuthResponse'
        "400":
          description: Invalid input or user not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Invalid credentials or user not verified
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Log in a user
      tags:
      - Authentication
...
  /api/password-reset:
    get:
      consumes:
      - application/json
      description: Sends a password reset code to the user's email.
      parameters:
      - description: User email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Password reset code sent successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Email is required or user not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Request a password reset
      tags:
      - Authentication
    post:
      consumes:
      - application/json
      description: Resets a user's password using the provided reset code and new
        password.
      parameters:
      - description: Password reset data
        in: body
        name: reset
        required: true
        schema:
          $ref: '#/definitions/controllers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input or user not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Invalid or expired reset code
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Reset a user's password
      tags:
      - Authentication
  /api/register:
    post:
      consumes:
      - application/json
      description: Registers a new user, sends a verification code to their email,
        and returns a token.
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/controllers.AuthResponse'
        "400":
          description: Invalid input or user already exists
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
...
  /api/verify:
    post:
      consumes:
      - application/json
      description: Verifies a user's email using the provided code and returns a token
        upon successful verification.
      parameters:
      - description: Verification data
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/controllers.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User verified successfully
          schema:
            $ref: '#/definitions/controllers.AuthResponse'
        "400":
          description: Invalid or missing data
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Invalid or expired verification code
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Verify a user's email
      tags:
      - Authentication
schemes:
- http
swagger: "2.0"